requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==4.2
django-ckeditor==6.7.1
django-cleanup==8.1.0
django-filter==24.2
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
packaging==24.0
pillow==10.2.0
PyJWT==1.7.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
